Update index.tsx
Added Loading component designed to show the user a spinner if the page they have selected is not loaded at that time.

Moved React based styling to logged out button page to fit in with AB testing results.

Removed Register button as Auth0 does not handle registration through a different api call.

Started creation of "dashboard" page to be displayed when the user is logged in.